Create parallel.cpp and parallel.h and move PV code into it
