+++ src/Makefile.am
@@ -122,6 +122,7 @@ BITCOIN_CORE_H = \
   leakybucket.h \
   netbase.h \
   noui.h \
+  parallel.h \
   policy/fees.h \
   policy/policy.h \
   policy/rbf.h \
@@ -208,6 +209,7 @@ libbitcoin_server_a_SOURCES = \
   miner.cpp \
   net.cpp \
   noui.cpp \
+  parallel.cpp \
   policy/fees.cpp \
   policy/policy.cpp \
   pow.cpp \
