+++ src/init.cpp
@@ -24,6 +24,7 @@
 #include "main.h"
 #include "miner.h"
 #include "net.h"
+#include "parallel.h"
 #include "policy/policy.h"
 #include "rpcserver.h"
 #include "script/standard.h"
@@ -168,7 +169,8 @@ static boost::scoped_ptr<ECCVerifyHandle> globalVerifyHandle;
 void Interrupt(boost::thread_group& threadGroup)
 {
     // Interrupt Parallel Block Validation threads if there are any running.
-    InterruptBlockValidationThreads();
+    PV.StopAllValidationThreads();
+    PV.WaitForAllValidationThreadsToStop();
 
     InterruptHTTPServer();
     InterruptHTTPRPC();
@@ -1128,7 +1130,7 @@ bool AppInit2(boost::thread_group& threadGroup, CScheduler& scheduler)
     LogPrintf("Using %u threads for script verification\n", nScriptCheckThreads);
     if (nScriptCheckThreads) {
         for (int i=0; i<nScriptCheckThreads-1; i++)
-            threadGroup.create_thread(&ThreadScriptCheck);
+            threadGroup.create_thread(&ThreadScriptCheck1);
     }
     // BU: parallel block validation - begin
     if (nScriptCheckThreads) {
