+++ src/test/script_P2SH_tests.cpp
@@ -5,7 +5,7 @@
 
 #include "key.h"
 #include "keystore.h"
-#include "main.h"
+#include "parallel.h"
 #include "policy/policy.h"
 #include "script/script.h"
 #include "script/script_error.h"
