+++ src/test/test_bitcoin.cpp
@@ -13,6 +13,7 @@
 #include "key.h"
 #include "main.h"
 #include "miner.h"
+#include "parallel.h"
 #include "pubkey.h"
 #include "random.h"
 #include "txdb.h"
@@ -73,7 +74,7 @@ TestingSetup::TestingSetup(const std::string& chainName) : BasicTestingSetup(cha
 
         nScriptCheckThreads = 3;
         for (int i=0; i < nScriptCheckThreads-1; i++)
-            threadGroup.create_thread(&ThreadScriptCheck);
+            threadGroup.create_thread(&ThreadScriptCheck1);
         for (int i=0; i<nScriptCheckThreads-1; i++)
             threadGroup.create_thread(&ThreadScriptCheck2);
         for (int i=0; i<nScriptCheckThreads-1; i++)
